OpenRouteservice webapp: do not build during Docker image build
The availablitliy of the dependencies of the web application is not
guarateed. This can lead to a broken web application. Therefore, the
fork of the web application contains a Docker setup to produce a static
build. This repository contains the result of the build as a simple
copy. NodeJS does not have to be executed in order to build the
openrouteservice-proxy container any more.